requirements.txt
streamlit>=1.35
streamlit-folium>=0.20
folium>=0.16
requests>=2.31
mysql-connector-python>=8.3
bcrypt>=4.1
gtts>=2.5
pandas>=2.0
altair>=5.0